Make registry a public package
