+++ internal/workflow/cache/cache_test.go
@@ -14,8 +14,8 @@ import (
 	"github.com/cschleiden/go-workflows/backend/metadata"
 	"github.com/cschleiden/go-workflows/core"
 	"github.com/cschleiden/go-workflows/internal/metrics"
-	"github.com/cschleiden/go-workflows/internal/registry"
 	wf "github.com/cschleiden/go-workflows/internal/workflow"
+	"github.com/cschleiden/go-workflows/registry"
 	"github.com/cschleiden/go-workflows/workflow"
 	"github.com/stretchr/testify/require"
 	"go.opentelemetry.io/otel/trace"
